requirements.txt
Django>=5.2.7
psycopg2-binary>=2.9.9
djangorestframework>=3.16.1
python-decouple>=3.8
djangorestframework-simplejwt>=5.3.0
pytest>=8.4.2
pytest-django>=4.11.1
pytest-cov>=7.0.0
pytest-mock>=3.15.1
